refactored types to individual files for compartmentalized impls
